Simplified project structure.
