+++ .env.example
@@ -1,13 +1,12 @@
 # See:
 # https://discord.com/developers/applications/
 PUBLIC_KEY=""
-GUILD_ID=""
 BOT_ID=""
 BOT_TOKEN=""
 
 # Optional environment variables
-PORT=8080
-DEV=1
+# PORT=8080
+# DEV=1
 
 # Perk provider environment variables
-OPENAI_API_KEY=""
+# OPENAI_API_KEY=""
