+++ README.md
@@ -10,19 +10,12 @@ Set of commands to mint, award, and use perks.
 
 ### Run
 
-1. Create a `.env` file with the following content:
-
-```env
-PUBLIC_KEY=""
-GUILD_ID=""
-BOT_TOKEN=""
-BOT_ID=""
-```
+1. Create a `.env` file based on `.env.example`.
 
 2. Run the bot server:
 
 ```bash
-deno run -Ar --unstable bot/http/main.ts
+deno run -Ar --unstable main.ts
 
 # Or
 
@@ -37,11 +30,7 @@ deno task start
 <img width="483" alt="image" src="https://user-images.githubusercontent.com/31261035/206064674-510f41f7-06c2-4899-8ace-b9451a8b0ad8.png">
 </a>
 
-Our development invite link:
-
-```url
-https://discord.com/api/oauth2/authorize?client_id=1048046283367133246&permissions=2733747599424&scope=bot
-```
+Development invite link: http://localhost:8080/__invite
 
 ### Update dependencies
